converter/pyproject.toml
[tool.poetry]
name = "hackathon_fair_python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
celery = "^5.2.7"
python-dotenv = "^0.20.0"
redis = "^4.3.4"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
django-mptt = "^0.13.4"
pika = "^1.3.0"
xmltodict = "^0.13.0"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
